+++ apm-sniffer/apm-sdk-plugin/jedis-plugins/jedis-2.x-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/jedis/v3/define/PipelineInstrumentation.java
@@ -66,7 +66,7 @@ public class PipelineInstrumentation extends AbstractWitnessInstrumentation {
                 new InstanceMethodsInterceptPoint() {
                     @Override
                     public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                        return named("sync").or(named("syncAndReturnAll"));
+                        return named("sync").or(named("syncAndReturnAll")).or(named("discard"));
                     }
 
                     @Override