+++ CHANGES.md
@@ -10,6 +10,7 @@ Release Notes.
 * Add plugin to trace Spring 6 and Resttemplate 6.
 * Move the baseline to JDK 17 for development, the runtime baseline is still Java 8 compatible.
 * Remove Powermock entirely from the test cases.
+* Refactor pipeline in jedis-plugin.
 
 #### Documentation
 * Update docs of Tracing APIs, reorganize the API docs into six parts